[runtime-security] extract important generator types to module
